+++ cypress/cypress/component/Input/Select.cy.tsx
@@ -89,3 +89,33 @@ it("option is disabled", () => {
 
   cy.get(`select[name=${name}]`).get('[value="DisabledOption"]').should("be.disabled");
 });
+
+it.only("undefined option value is working", () => {
+  const name = faker.random.word();
+  const schema = yup.object().shape({
+    [name]: yup.string(),
+  });
+  const options = faker.helpers.uniqueArray<LabelValueOption>(() => ({ value: faker.random.alpha(10), label: faker.random.alpha(10) }), 3);
+
+  options[0].value = undefined;
+
+  cy.mount(
+    <Form onSubmit={cy.spy().as("onSubmitSpy")} resolver={yupResolver(schema)}>
+      <Input type="select" name={name} label={name} options={options} />
+      <input type={"submit"} />
+    </Form>,
+  );
+
+  cy.get("input[type=submit]").click({ force: true });
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: undefined });
+
+  cy.get("@onSubmitSpy").invoke("resetHistory");
+  cy.get(`select[id=${name}]`).select(options[1].label);
+  cy.get("input[type=submit]").click({ force: true });
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: options[1].value });
+
+  cy.get("@onSubmitSpy").invoke("resetHistory");
+  cy.get(`select[id=${name}]`).select(options[0].label);
+  cy.get("input[type=submit]").click({ force: true });
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: undefined });
+});